Extract mentionTag helpers into a separate file
